+++ lib/web/ui/mfa.go
@@ -17,11 +17,7 @@ package ui
 import (
 	"time"
 
-	"github.com/gravitational/teleport/api/client/proto"
 	"github.com/gravitational/teleport/api/types"
-	"github.com/gravitational/teleport/lib/auth"
-	"github.com/gravitational/teleport/lib/auth/u2f"
-	wanlib "github.com/gravitational/teleport/lib/auth/webauthn"
 )
 
 // MFADevice describes a mfa device
@@ -54,30 +50,4 @@ func MakeMFADevices(devices []*types.MFADevice) []MFADevice {
 	}
 
 	return uiList
-}
-
-// MakeAuthenticateChallenge converts proto to JSON format.
-func MakeAuthenticateChallenge(protoChal *proto.MFAAuthenticateChallenge) *auth.MFAAuthenticateChallenge {
-	chal := &auth.MFAAuthenticateChallenge{
-		TOTPChallenge: protoChal.GetTOTP() != nil,
-	}
-
-	for _, u2fChal := range protoChal.GetU2F() {
-		ch := u2f.AuthenticateChallenge{
-			Version:   u2fChal.Version,
-			Challenge: u2fChal.Challenge,
-			KeyHandle: u2fChal.KeyHandle,
-			AppID:     u2fChal.AppID,
-		}
-		if chal.AuthenticateChallenge == nil {
-			chal.AuthenticateChallenge = &ch
-		}
-		chal.U2FChallenges = append(chal.U2FChallenges, ch)
-	}
-
-	if protoChal.GetWebauthnChallenge() != nil {
-		chal.WebauthnChallenge = wanlib.CredentialAssertionFromProto(protoChal.WebauthnChallenge)
-	}
-
-	return chal
 }
